% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_in_modifiers.R
\name{adjust_coeffs}
\alias{adjust_coeffs}
\title{adjust the two elements of the list for the markov wx generator}
\usage{
adjust_coeffs(coeffs, data, mean_mult = 1, adjust_sd = TRUE)
}
\arguments{
\item{coeffs}{list created by rSOILWAT2::dbW_estimate_WGen_coefs}

\item{data}{daily wather data (dataframe)}

\item{mean_mult}{what to multiply mean event sizes by}

\item{adjust_sd}{logical, whether to adjust the PPT_sd for a give DOY.
This option is used so that a draw from a normal distribution with a new
(larger) mean is no more/less likely draw a negative value}
}
\value{
a list with same elements as
}
\description{
The purpose of this function is to adjust the wgen input files
so that simulated weather has an increased mean event size,
decreased ppt frequency, and no change in mean temperature. Also no change
in total precipitation
}
\examples{
data <-data.frame(rSOILWAT2::dbW_weatherData_to_dataframe(rSOILWAT2::weatherData))
coeffs <- rSOILWAT2::dbW_estimate_WGen_coefs(data)

}
