% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_in_modifiers.R
\name{adjust_covar_in}
\alias{adjust_covar_in}
\title{adjust weekly correction temp correction factors}
\usage{
adjust_covar_in(covar_in, data, mean_mult = 1)
}
\arguments{
\item{covar_in}{dataframe of weekly temp datagenerated by
rSOILWAT2::dbW_weatherData_to_dataframe}

\item{data}{dataframe of daily temp and precip}

\item{mean_mult}{multiplier that mean precipitation event size is to
be multiplied by}
}
\value{
dataframe with same columns as covar_in, except the CF have been
adjusted (if mean_mult not 1)
}
\description{
adjust weekly correction temp correction factors
}
\examples{
data <-data.frame(rSOILWAT2::dbW_weatherData_to_dataframe(rSOILWAT2::weatherData))
covar_in <- rSOILWAT2::dbW_estimate_WGen_coefs(data)[[1]]
head(adjust_covar_in(covar_in, data))
head(adjust_covar_in(covar_in, data, mean_mult = 2))
}
