% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_in_modifiers.R
\name{adjust_mkv_woy}
\alias{adjust_mkv_woy}
\title{adjust weekly correction temp correction factors}
\usage{
adjust_mkv_woy(mkv_woy, data, mean_mult = 1)
}
\arguments{
\item{mkv_woy}{dataframe of weekly temp datagenerated by
rSOILWAT2::dbW_weatherData_to_dataframe}

\item{data}{dataframe of daily temp and precip}

\item{mean_mult}{multiplier that mean precipitation event size is to
be multiplied by}
}
\value{
dataframe with same columns as mkv_woy, except the CF have been
adjusted (if mean_mult not 1)
}
\description{
adjust weekly correction temp correction factors
}
\examples{
data <-data.frame(rSOILWAT2::dbW_weatherData_to_dataframe(rSOILWAT2::weatherData))
mkv_woy <- rSOILWAT2::dbW_estimate_WGen_coefs(data)[[1]]
head(adjust_mkv_woy(mkv_woy, data))
head(adjust_mkv_woy(mkv_woy, data, mean_mult = 2))
}
