% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGEN_ppt.R
\name{adjust_params}
\alias{adjust_params}
\title{Adjust parameters of monthly precipitation}
\usage{
adjust_params(params, mean_mult, sd_mult)
}
\arguments{
\item{params}{dataframe of parameters from monthly_ppt_params() function}

\item{mean_mult}{amount to multiply the daily mean by}

\item{sd_mult}{amount to multiply the standard deviation by}
}
\value{
adjusted parameters with probability of rain multiplied by
a factor of 1/mean_mult so that total precip is held constant
}
\description{
Adjust parameters of monthly precipitation
}
\examples{
mean_mult <- 1.5
sd_mult <- 2
params <- monthly_ppt_params(wx_data)
params2 <- adjust_params(params, mean_mult, sd_mult)
df <- markov_chain(params2)
df2 <- generate_events(df, params2)
# make histograms
if (FALSE){
breaks <- seq(from = 0, to =10, by = 0.25)
hist(df2$PPT_cm[df2$PPT_cm > 0], breaks = breaks)
hist(wx_data$PPT_cm[wx_data$PPT_cm > 0],breaks = breaks)
}
}
