% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGEN_temp_modifiers.R
\name{adjust_temp}
\alias{adjust_temp}
\title{adjust temperature for wet and dry days}
\usage{
adjust_temp(ppt_df, temp_df, wk_list, mean_mult = 1)
}
\arguments{
\item{ppt_df}{data frame of precip data (from generate_events())}

\item{temp_df}{dataframe of temp data from generate_temp()}

\item{wk_list}{list of weakly temperature parameters from temp_wk_list()}

\item{mean_mult}{scalar that mean is multiplied when increasing ppt intensity}
}
\value{
dataframe of PPT and temp, with Tmax_C and Tmin_C adjusted
for wet/dry days
}
\description{
adjust temperature for wet and dry days
}
\examples{
params <- monthly_ppt_params(wx_data)
df1 <- markov_chain(params,  start_date = "1980-01-01", end_date = "1980-12-31")
ppt_df <- generate_events(df1, params)
wk_list <- temp_wk_list(wx_data)
temp_df <- generate_temp(wk_list, start_date = "1980-01-01", end_date = "1980-12-31")
comb <- adjust_temp(ppt_df, temp_df, wk_list)
mean(comb$Tmax_C); mean(temp_df$Tmax_C)
}
