% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGEN_temp.R
\name{calc_A}
\alias{calc_A}
\title{Calculate A matrix}
\usage{
calc_A(M0, M1)
}
\arguments{
\item{M0}{M0 matrix (from calc_M0() function)}

\item{M1}{M1 matrix (from calc_M1() function)}
}
\value{
2x2 matrix
}
\description{
Equation 4 in Richardson and Wright (1984)
}
\examples{
df <- dplyr::filter(wx_data, month == 2 & day \%in\% 1:7)
Tmax <- df$Tmax_C
Tmin <- df$Tmin_C
year <- df$year
M0 <- calc_M0(Tmax, Tmin)
M1 <- calc_M1(Tmax, Tmin, year)
A <- calc_A(M0, M1)
}
