% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGEN_temp.R
\name{calc_B}
\alias{calc_B}
\title{Calculate B matrix}
\usage{
calc_B(M0, M1)
}
\arguments{
\item{M0}{M0 matrix (from calc_M0() function)}

\item{M1}{M1 matrix (from calc_M1() function)}
}
\value{
2x2 matrix
}
\description{
Equation 5 in Richardson and Wright (1984)
}
