% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGEN_temp.R
\name{calc_M0}
\alias{calc_M0}
\title{calculate the M0 matrix}
\usage{
calc_M0(Tmax, Tmin)
}
\arguments{
\item{Tmax}{daily max temperature vector}

\item{Tmin}{daily min temperature vector}
}
\value{
2x2 matrix
}
\description{
This is the autocorrelation/correlation matrix described in
eq 8 in Richardson and Wright 1984, and is similar to Wilks, 2011, eq 11.21,
except that Wilks uses covariance instead
}
\examples{
df <- dplyr::filter(wx_data, month == 1 & day \%in\% 1:7)
year <- df$year
Tmax <- df$Tmax_C
Tmin <- df$Tmin_C
}
