% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGEN_temp.R
\name{calc_M1}
\alias{calc_M1}
\title{Calculate the M1 matrix}
\usage{
calc_M1(Tmax, Tmin, year)
}
\arguments{
\item{Tmax}{daily max temperature vector}

\item{Tmin}{daily min temperature vector}

\item{year}{vector of year. Year is included so lag 1 one works (not finding correlation
with lagged value of the previous year)}
}
\value{
2x2 matrix
}
\description{
Lag 1 autocorrelation and correlation of min and max temp.
eq 11.22b in Wilks 2011, except wilks used covariance instead of correlation.
This is eq. 7 in Richardson and Wright (1984).
Since I am calculating the
lagged correlation by 'hand' (not using acf())
the values are slightly (1% for one data set I check)
different (b/ of how means calculated), but I'm
doing that to deal with the issue of lagging across years.
}
