% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_dfs.R
\name{calc_map}
\alias{calc_map}
\title{calculate mean annual precipitation}
\usage{
calc_map(wout)
}
\arguments{
\item{wout}{either a dataframe of rSOILWAT2 weather list}
}
\value{
single value (MAP)
}
\description{
calculate mean annual precipitation
}
