% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_dfs.R
\name{calc_mat}
\alias{calc_mat}
\title{calculate mean temperature}
\usage{
calc_mat(wout, col)
}
\arguments{
\item{wout}{either a dataframe of rSOILWAT2 weather list}

\item{col}{string, name of the column to take an average of}
}
\value{
single value (MAT)
}
\description{
calculate mean temperature
}
\examples{
calc_mat(rSOILWAT2::weatherData, col = "Tmax_C")
}
