% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_dfs.R
\name{calc_nwet}
\alias{calc_nwet}
\title{calculate average number of wet days per year}
\usage{
calc_nwet(wout)
}
\arguments{
\item{wout}{either a dataframe of rSOILWAT2 weather list}
}
\value{

}
\description{
calculate average number of wet days per year
}
