% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_in_modifiers.R
\name{calc_p_W}
\alias{calc_p_W}
\title{probability of rain each day based on mkv_doy}
\usage{
calc_p_W(mkv_doy)
}
\arguments{
\item{mkv_doy}{dataframe of daily markov ppt transition probabilities}
}
\value{
numeric vector giving probability of rain each day (non conditional)
}
\description{
probability of rain each day based on mkv_doy
}
\examples{
data <-data.frame(rSOILWAT2::dbW_weatherData_to_dataframe(rSOILWAT2::weatherData))
prob_in <- rSOILWAT2::dbW_estimate_WGen_coefs(data)[[2]]
calc_p_W(mkv_doy = prob_in)
}
