% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_intensity.R
\name{create_event_df}
\alias{create_event_df}
\title{Create dataframe with event numbers}
\usage{
create_event_df(x)
}
\arguments{
\item{x}{numeric vector (daily precip)}
}
\value{
tibble with 2 columns, x (input vector), and event which is the
event number corresponding to that value of x
}
\description{
Create dataframe with event numbers
}
\examples{
create_event_df(c(0.1, 0.1, 0, 0.2))
}
