% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_in_modifiers.R
\name{expected_nwet}
\alias{expected_nwet}
\title{expected number of wet days}
\usage{
expected_nwet(coeffs, adjust_for_truncnorm = FALSE)
}
\arguments{
\item{coeffs}{list that contains mkv_doy element}

\item{adjust_for_truncnorm}{logical, whether to adjust the number of
expected wet days based on fact that the weather generator uses a truncated
normal distribution (i.e. some 'wet' days then become dry)}
}
\value{
expected number of wet days per year
}
\description{
expected number of wet days
}
