% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_in_modifiers.R
\name{expected_ppt}
\alias{expected_ppt}
\title{expected annual ppt based on markov coefficients}
\usage{
expected_ppt(coeffs, adjust_for_truncnorm = FALSE)
}
\arguments{
\item{coeffs}{--list that contains mkv_doy element}

\item{adjust_for_truncnorm}{logical, whether to adjust the number of
expected wet days based on fact that the weather generator uses a truncated
normal distribution (i.e. some 'wet' days then become dry)}
}
\value{
expected value of annual precipitation
}
\description{
Calculates probability of rain each day and from that calculates
expected value of annual precip. Accounts for the fact that when wgen draws
negative value from normal distribution that day is converted to a dry day--
this is still in testing, and currently I have removed this feature (ie not
accounting for truncated normal???)
The problem with this function currently is that while observed/simulated
 and expected
ppt is normally very close, when the sd is increased, the mismatch grows
}
\examples{
data <-data.frame(rSOILWAT2::dbW_weatherData_to_dataframe(rSOILWAT2::weatherData))
coeffs <- rSOILWAT2::dbW_estimate_WGen_coefs(data)
ex <- expected_ppt(coeffs, adjust_for_truncnorm = TRUE)
ex
}
