% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGEN_ppt.R
\name{generate_events}
\alias{generate_events}
\title{generate precipitation events}
\usage{
generate_events(df, params)
}
\arguments{
\item{df}{dataframe, output from markov_chain() function}

\item{params}{dataframe of parameters from monthly_ppt_params() function
or from adjust_params() function}
}
\value{
dataframe with PPT_cm (daily precip) column added
}
\description{
Generate daily precipitation on wet days based on gamma distribution
described by mean and SD in the params dataframe
}
\examples{
params <- monthly_ppt_params(wx_data)
df <- markov_chain(params)
df2 <- generate_events(df, params)
# calculate yearly ppt (original data is 31.6 cm)
df2 \%>\% dplyr::group_by(year) \%>\%
dplyr::summarize(PPT_cm = sum(PPT_cm)) \%>\%
dplyr::summarize(PPT_cm = mean(PPT_cm))
}
