% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGEN_temp.R
\name{generate_temp}
\alias{generate_temp}
\title{generate temperature}
\usage{
generate_temp(wk_list, start_date = "1980-01-01", end_date = "2010-12-31")
}
\arguments{
\item{wk_list}{list of weekly parameters from temp_wk_list() function}

\item{start_date}{"yyy-mm-dd" string for start date of simulated temperature}

\item{end_date}{"yyy-mm-dd" string for end date of simulated temperature}
}
\value{
dataframe, that includes simulated minimum and maximum temperature
}
\description{
generate temperature
}
\examples{
wk_list <- temp_wk_list(wx_data)
test <- generate_temp(wk_list, "1980-01-01", "1982-12-31")
plot(test$Tmax_C, col = "red", type = "l")
lines(test$Tmin_C, col = "blue")
hist((test$Tmax_C-test$Tmin_C))
# testing
if (FALSE) {
test <- generate_temp(wk_list, "1980-01-01", "2079-12-31") # 30 sec to run
dplyr::summarize(test, Tmax_C = mean(Tmax_C), Tmin_C = mean(Tmin_C))
# compare to original data
dplyr::summarize(wx_data, Tmax_C = mean(Tmax_C), Tmin_C = mean(Tmin_C))
}
}
