% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/increase_intensity.R
\name{incr_dly_intensity}
\alias{incr_dly_intensity}
\title{Increase daily precipitation intensity}
\usage{
incr_dly_intensity(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric vector with same. Doubles daily precipitation intensity.
Takes precip from odd days and adds it to even days.
}
\description{
Increase daily precipitation intensity
}
\examples{
x <- c(0, 0, 0, 0.1, 0, 2, 0.1, 0, 0, 1, 0, 0.2, 0.2, 0)
x2 <- incr_dly_intensity(x)
x2
}
