% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/increase_intensity.R
\name{incr_event_intensity}
\alias{incr_event_intensity}
\title{Increase event size intensity}
\usage{
incr_event_intensity(x, from = 1, to = 1)
}
\arguments{
\item{x}{numeric vector (daily precip)}

\item{from}{positive integer. The number of events to take precip from
and add to the following event(s)}

\item{to}{positive integer. The number of events to add the removed precip
to}
}
\value{
numeric vector, same length as x. Precip events are considered
1 or more consecutive days with precip. Precip was removed from one or more (from)
events and added to one or more (to) events, and this is repeated for all events
}
\description{
Increase event size intensity
}
\examples{
y1 <- incr_event_intensity(c(0.1, 0.1, 0, 1))
all.equal(y1, c(0, 0, 0, 1.2))
y2 <- incr_event_intensity(c(0.2, 0.1, 0.1, 0, 1, 2))
all.equal(y2, c(0, 0, 0, 0, 1.2, 2.2))
# 2 events added to 1
y3 <- incr_event_intensity(c(0.1, 0, 0.2, 0.1, 0, 1),
                           from = 2, to = 1)
all.equal(y3, c(0, 0, 0, 0, 0, 1.4))
y4 <- incr_event_intensity(c(0.3, 0, 0.2, 0.1, 0, 1, 0, 1),
                           from = 1, to = 2)
all.equal(y4, c(0, 0, 0.3, 0.2, 0, 1.1, 0, 1))
}
