% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGEN_temp.R
\name{join_temp_ppt}
\alias{join_temp_ppt}
\title{join temp and precip dataframes}
\usage{
join_temp_ppt(ppt_df, temp_df)
}
\arguments{
\item{ppt_df}{dataframe with precip (must have "date" column)}

\item{temp_df}{dataframe with temperature (must have "date" column)}
}
\value{
dataframe joined on "date"
}
\description{
join temp and precip dataframes
}
\examples{
params <- monthly_ppt_params(wx_data)
df1 <- markov_chain(params,  start_date = "1980-01-01", end_date = "1980-12-31")
ppt_df <- generate_events(df1, params)
wk_list <- temp_wk_list(wx_data)
temp_df <- generate_temp(wk_list, start_date = "1980-01-01", end_date = "1980-12-31")
}
