% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGEN_ppt.R
\name{markov_chain}
\alias{markov_chain}
\title{first order markov chain}
\usage{
markov_chain(params, start_date = "1980-01-01", end_date = "2010-12-31")
}
\arguments{
\item{params}{dataframe of parameters from monthly_ppt_params() function or
from adjust_params() function}

\item{start_date}{start date of simulation}

\item{end_date}{end date of simulation}
}
\value{
dataframe, with date, year, month, day and is_wet columns. This final
column is logical of whether precip occured that day
}
\description{
Markov chain used simulate wet and dry days
}
\examples{
params <- monthly_ppt_params(wx_data)
df <- markov_chain(params)
df
sum(df$is_wet)/nrow(df) # percent of days that are wet
sum(wx_data$PPT_cm > 0)/nrow(wx_data) # comparison to original data
}
