% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_intensity.R
\name{max_event_length}
\alias{max_event_length}
\title{Max rain event length}
\usage{
max_event_length(x)
}
\arguments{
\item{x}{numeric vector of daily rain events}
}
\value{
the length of the longest sequence of consecutive rainy days
}
\description{
Max rain event length
}
\examples{
max_event_length(c(0, 1, 1, 2, 0, 1)) == 3
}
