% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_intensity.R
\name{mean_event_size}
\alias{mean_event_size}
\title{Calculate mean event size}
\usage{
mean_event_size(x)
}
\arguments{
\item{x}{numeric vector (daily precip)}
}
\value{
mean event size, where an event is the amount of rain on one or more
consecutive days of rain
}
\description{
Calculate mean event size
}
\examples{
mean_event_size(c(0.1, 0.1, 0, 0.2))
mean_event_size(0)
mean_event_size(rep(0.1, 10))
}
