% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGEN_ppt.R
\name{monthly_ppt_params}
\alias{monthly_ppt_params}
\title{monthly precipitation parameters}
\usage{
monthly_ppt_params(data)
}
\arguments{
\item{data}{dataframe of daily precipitation  (must include PPT_cm, year,
and  month columns)}
}
\value{
dataframe
}
\description{
monthly precipitation parameters
}
\examples{
monthly_ppt_params(wx_data)
}
