% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_intensity.R
\name{n_events}
\alias{n_events}
\title{Number of rain events}
\usage{
n_events(x, min_length = 1)
}
\arguments{
\item{x}{numeric vector}

\item{min_length}{a positive integer. The minimum number of days an event
has to be in order to count it.}
}
\value{
The number of rain events in x. Rain events are defined
as a sequence with 1 (default) or more consecutive days with rain
}
\description{
Number of rain events
}
\examples{
n_events(rep(c(0, 1), 5)) == 5
n_events(rep(c(0, 1), each = 5)) == 1
n_events(c(1, 1, 0, 0.5, 0, .1, .2, 0), min_length = 2) == 2
}
