% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGEN_ppt.R
\name{P_W_X}
\alias{P_W_X}
\title{probability of rain given previous day's wet/dry status}
\usage{
P_W_X(x, year, return = "both")
}
\arguments{
\item{x}{numeric vector of daily precip}

\item{year}{numeric vector. Passing function the year as well,
so don't 'look' at previous day that is actually in previous year.}

\item{return}{use "both" (return both probabilities), "P_W_W" (probability
is wet given previous day was wet), or "P_W_D" (prob wet given previous
day was dry)}
}
\value{
vector of length 1 or 2
}
\description{
probability of rain given previous day's wet/dry status
}
\examples{
x <- wx_data$PPT_cm[wx_data$month == 1]
year <- wx_data$year[wx_data$month == 1]
return <- "both"
P_W_X(x, year, return)
}
