% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGEN_misc.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe}
\arguments{
\item{lhs, rhs}{see magrittr documentation}
}
\description{
Re-exporting the magrittr, \code{\%>\%}. I do this
because the pipe is used in code examples in the package
}
