% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGEN_temp_modifiers.R
\name{temp_modifier}
\alias{temp_modifier}
\title{temperature modifier}
\usage{
temp_modifier(all_days, x_days)
}
\arguments{
\item{all_days}{mean temp on all days}

\item{x_days}{mean temp on wet (dry) days}
}
\value{
scalar modifier for wet (dry) days
}
\description{
Lower (raise) temperature of a day based on whether it is wet
(dry).
}
\examples{
temp_modifier(10, 2)
}
