% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGEN_temp.R
\name{temp_wk_list}
\alias{temp_wk_list}
\title{List of weekly temperature parameters}
\usage{
temp_wk_list(data)
}
\arguments{
\item{data}{dataframe of original weather dat with "date", "Tmax_C",
"Tmin_C", "PPT_cm" columns}
}
\value{
list with 52 elements (1 for each week of the year, week 52 and 53
are combined). Each list element is a list that contains several elements
including: "mean", mean temperature on all days; "sd", sd of temp;
"wet_dry_temp", which is mean temp and wet and dry days; "A", A matrix;
"B", B matrix.
}
\description{
List of weekly temperature parameters
}
\examples{
wk_list <- temp_wk_list(wx_data)
# plotting min/max temps
Tmax_mean <- purrr::map_dbl(wk_list, function(x) x$means[1])
plot(Tmax_mean, type = "l", col = "red")
lines(purrr::map_dbl(wk_list, function(x) x$means[2]), col = "blue")
}
