% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_intensity.R
\name{wet_lengths}
\alias{wet_lengths}
\title{Lengths of rain events}
\usage{
wet_lengths(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
vector containing the lengths (number of consecutive days) of
rain events
}
\description{
Utility function used in subsequent functions
}
\examples{
wet_lengths(c(0, 1, 1, 0, 1))
wet_lengths(c(0, 0))
}
