% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGEN_temp.R
\name{wk_summary_wgen}
\alias{wk_summary_wgen}
\title{dataframe of weekly temp parameters for adjusting rSOILWAT Wgen coeffs}
\usage{
wk_summary_wgen(data, mean_mult = 1)
}
\arguments{
\item{data}{dataframe of original weather dat with "Year", "DOY", "Tmax_C",
"Tmin_C", "PPT_cm" columns (columns as needed by
rSOILWAT2::dbW_estimate_WGen_coefs)}
}
\value{
a dataframe with mean temperature, and number of wet days in each
week of year
}
\description{
dataframe of weekly temp parameters for adjusting rSOILWAT Wgen coeffs
}
\examples{
data <-data.frame(rSOILWAT2::dbW_weatherData_to_dataframe(rSOILWAT2::weatherData))
head(wk_summary_wgen(data))
head(rSOILWAT2::dbW_estimate_WGen_coefs(data)[[1]])
# compare
wk_summary_wgen(data, mean_mult = 2)
}
