% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{wx_data}
\alias{wx_data}
\title{Example dataset of daily temperature and precipitation}
\format{
A data frame with 11223 rows and 9 variables:
\describe{
  \item{site}{site number, just one site in this dataset}
  \item{year}{year of measurement}
  \item{month}{month of measurement}
  \item{day}{day of month}
  \item{date}{date, yyyy-mm-dd}
  \item{DOY}{day of year}
  \item{Tmax_C}{daily max temp, C}
  \item{Tmin_C}{daily min temp, C}
  \item{PPT_cm}{daily precipitation, cm}
}
}
\usage{
wx_data
}
\description{
A dataset for daily weather (from daymet 1 km product). This is the daily
weather for site 5 in Palmquist et al in press (2021).
}
\keyword{datasets}
