% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGEN_ppt.R
\name{P_W_X}
\alias{P_W_X}
\title{probability of rain given previous day's wet/dry status}
\usage{
P_W_X(x, return = "both")
}
\arguments{
\item{x}{numeric vector of daily precip}

\item{return}{use "both" (return both probabilities), "P_W_W" (probability
is wet given previous day was wet), or "P_W_D" (prob wet given previous
day was dry)}
}
\value{
vector of length 1 or 2
}
\description{
Note: Next step is improve function by passing it the year as well,
so don't 'look' at previous day that is actually the previous year
}
\examples{
x <- wx_data$PPT_cm[wx_data$month == 1]
P_W_X(x)
}
